pkg up nerdstats and profiler bits.
